+++ app/admin/visits.rb
@@ -98,20 +98,18 @@ ActiveAdmin.register Visit do
     column :description
     column :visit_type
     column :visit_date, sortable: false
-    if(can? :read_qc, Visit or (not Rails.application.config.is_erica_remote))
-      column :state, :sortable => :state do |visit|
-        next unless (can? :read_qc, visit or (not Rails.application.config.is_erica_remote))
-
-        case(visit.state_sym)
-        when :incomplete_na then status_tag('Incomplete, not available')
-        when :complete_tqc_passed then status_tag('Complete, tQC of all series passed', :ok)
-        when :incomplete_queried then status_tag('Incomplete, queried', :warning)
-        when :complete_tqc_pending then status_tag('Complete, tQC not finished', :warning)
-        when :complete_tqc_issues then status_tag('Complete, tQC finished, not all series passed', :error)
-        end
+    column :state, :sortable => :state do |visit|
+      case(visit.state_sym)
+      when :incomplete_na then status_tag('Incomplete, not available')
+      when :complete_tqc_passed then status_tag('Complete, tQC of all series passed', :ok)
+      when :incomplete_queried then status_tag('Incomplete, queried', :warning)
+      when :complete_tqc_pending then status_tag('Complete, tQC not finished', :warning)
+      when :complete_tqc_issues then status_tag('Complete, tQC finished, not all series passed', :error)
       end
+    end
+    if can? :read_mqc, Visit
       column 'mQC State', :mqc_state, :sortable => :mqc_state do |visit|
-        next unless (can? :read_qc, visit or (not Rails.application.config.is_erica_remote))
+        next unless can? :read_mqc, visit
 
         case(visit.mqc_state_sym)
         when :pending then status_tag('Pending')
@@ -120,14 +118,15 @@ ActiveAdmin.register Visit do
         end
       end
       column 'mQC Date', :mqc_date do |visit|
-        next unless (can? :read_qc, visit or (not Rails.application.config.is_erica_remote))
+        next unless can? :read_mqc, visit
 
         pretty_format(visit.mqc_date)
       end
       column 'mQC User', :mqc_user, :sortable => :mqc_user_id do |visit|
-        next unless (can? :read_qc, visit or (not Rails.application.config.is_erica_remote))
+        next unless can? :read_mqc, visit
+        next if visit.mqc_user.nil?
 
-        link_to(visit.mqc_user.name, admin_user_path(visit.mqc_user)) unless visit.mqc_user.nil?
+        link_to(visit.mqc_user.name, admin_user_path(visit.mqc_user))
       end
     end
     keywords_column(:tags, 'Keywords') if Rails.application.config.is_erica_remote
@@ -142,16 +141,16 @@ ActiveAdmin.register Visit do
       row :description
       row :visit_type
       row :visit_date
-      if(can? :read_qc, visit or (not Rails.application.config.is_erica_remote))
-        row :state do
-          case(visit.state_sym)
-          when :incomplete_na then status_tag('Incomplete, not available')
-          when :complete_tqc_passed then status_tag('Complete, tQC of all series passed', :ok)
-          when :incomplete_queried then status_tag('Incomplete, queried', :warning)
-          when :complete_tqc_pending then status_tag('Complete, tQC not finished', :warning)
-          when :complete_tqc_issues then status_tag('Complete, tQC finished, not all series passed', :error)
-          end
+      row :state do
+        case(visit.state_sym)
+        when :incomplete_na then status_tag('Incomplete, not available')
+        when :complete_tqc_passed then status_tag('Complete, tQC of all series passed', :ok)
+        when :incomplete_queried then status_tag('Incomplete, queried', :warning)
+        when :complete_tqc_pending then status_tag('Complete, tQC not finished', :warning)
+        when :complete_tqc_issues then status_tag('Complete, tQC finished, not all series passed', :error)
         end
+      end
+      if can? :read_mqc, visit
         row 'mQC State' do
           case(visit.mqc_state_sym)
           when :pending then status_tag('Pending')
@@ -260,11 +259,11 @@ ActiveAdmin.register Visit do
     column :created_at
     column :updated_at
     column :description
-    column('State') {|v| (can? :read_qc, v or (not Rails.application.config.is_erica_remote)) ? v.state_sym : ''}
-    column('Mqc Date') {|v| (can? :read_qc, v or (not Rails.application.config.is_erica_remote)) ? v.mqc_date : ''}
-    column('Mqc State') {|v| (can? :read_qc, v or (not Rails.application.config.is_erica_remote)) ? v.mqc_state_sym : ''}
-    column('Patient') {|v| v.patient.nil? ? '' : v.patient.name}
-    column('Visit Date') {|v| v.visit_date}
+    column('State') { |v| can?(:read_mqc, v) ? v.state_sym : '' }
+    column('Mqc Date') { |v| can?(:read_mqc, v) ? v.mqc_date : '' }
+    column('Mqc State') { |v| can?(:read_mqc, v) ? v.mqc_state_sym : '' }
+    column('Patient') { |v| v.patient.nil? ? '' : v.patient.name }
+    column('Visit Date') { |v| v.visit_date }
   end
 
   # filters
@@ -406,11 +405,7 @@ ActiveAdmin.register Visit do
 
   member_action :mqc_results, :method => :get do
     @visit = Visit.find(params[:id])
-    if(Rails.application.config.is_erica_remote)
-      authorize! :read_qc, @visit
-    else
-      authorize! :mqc, @visit unless can? :manage, @visit
-    end
+    authorize! :read_mqc, @visit
 
     mqc_version = @visit.mqc_version
     @mqc_spec = @visit.mqc_spec_with_results_at_version(mqc_version)
@@ -425,7 +420,7 @@ ActiveAdmin.register Visit do
   end
   member_action :mqc, :method => :post do
     @visit = Visit.find(params[:id])
-    authorize! :mqc, @visit unless can? :manage, @visit
+    authorize! :perform_mqc, @visit
 
     unless([:complete_tqc_passed, :complete_tqc_issues, :incomplete_na].include?(@visit.state_sym))
       flash[:error] = 'mQC cannot be performed for a visit in this state.'
@@ -450,7 +445,7 @@ ActiveAdmin.register Visit do
   end
   member_action :mqc_form, :method => :get do
     @visit = Visit.find(params[:id])
-    authorize! :mqc, @visit unless can? :manage, @visit
+    authorize! :perform_mqc, @visit
 
     unless([:complete_tqc_passed, :complete_tqc_issues, :incomplete_na].include?(@visit.state_sym))
       flash[:error] = 'mQC cannot be performed for a visit in this state.'
@@ -468,14 +463,12 @@ ActiveAdmin.register Visit do
     @page_title = "Perform mQC"
     render 'admin/visits/mqc_form'
   end
-  action_item :edit, :only => :show do
-    link_to('Perform mQC', mqc_form_admin_visit_path(resource)) if([:complete_tqc_passed, :complete_tqc_issues, :incomplete_na].include?(resource.state_sym) and (can? :mqc, resource or can? :manage, resource))
+  action_item :edit, only: :show, if: -> { resource.ready_for_mqc? && can?(:perform_mqc, resource) } do
+    link_to('Perform mQC', mqc_form_admin_visit_path(resource))
   end
-  action_item :edit, :only => :show do
-    link_to('mQC Results', mqc_results_admin_visit_path(resource)) if(resource.mqc_state_sym != :pending and (
-                                                                        (Rails.application.config.is_erica_remote and can? :read_qc, resource) or
-                                                                        (can? :mqc, resource or can? :manage, resource))
-                                                                     )
+
+  action_item :edit, only: :show, if: -> { resource.has_mqc_results? && can?(:read_mqc, resource) } do
+    link_to('mQC Results', mqc_results_admin_visit_path(resource))
   end
 
   member_action :edit_state, :method => :post do
@@ -491,6 +484,7 @@ ActiveAdmin.register Visit do
     end
     redirect_to params[:return_url]
   end
+
   member_action :edit_state_form, :method => :get do
     @visit = Visit.find(params[:id])
     authorize! :update_state, @visit